refactor: ReadPostDialog, AddCommentDialog 분리 ing
