+++ src/features/dialog/model/useShowEditPostDialogStore.ts
@@ -5,7 +5,7 @@ interface ShowEditDialogStoreState {
   setShowEditDialog: (show: boolean) => void
 }
 
-// 게시물 추가 대화상자 띄우기
+// 게시물 수정 대화상자 띄우기
 export const useShowEditDialogStore = create<ShowEditDialogStoreState>((set) => ({
   showEditDialog: false,
   setShowEditDialog: (show: boolean) => set(() => ({ showEditDialog: show })),
